twitter_bot/requirements.txt
dj-database-url==0.5.0
Django==3.1.1
gunicorn==20.0.4
whitenoise==5.2.0
tweepy==3.10.0
django-extensions==3.1.1
djangorestframework==3.12.4
request==2.26.0
dropbox